+++ packages/strategy/CHANGELOG.md
@@ -1,4 +1,4 @@
-## @remix-pwa/strategy 2.1.6 (2023-09-26)
+## @remix-pwa/strategy 2.1.6-dev.2 (2023-09-27)
 
 
 
@@ -6,7 +6,7 @@
 
 ### Dependencies
 
-* **@remix-pwa/cache:** upgraded to 2.0.9
+* **@remix-pwa/cache:** upgraded to 2.0.9-dev.3
 
 ## @remix-pwa/strategy 2.1.6-dev.1 (2023-09-26)